requirements.txt
fastapi[all]
jinja2
uvicorn
gunicorn
google-cloud
google-cloud-storage